requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
matplotlib==3.9.2
reportlab==4.2.5
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0